refactoring pages folder structure & added edit/delete review functionality
