+++ server/src/index.ts
@@ -43,6 +43,7 @@ app.get("/reviews/:id", queries.Reviews.getReviewById);
 app.post("/createCategory", jwtCheck, queries.Reviews.createCategory);
 app.post("/review/createReview", jwtCheck, queries.Reviews.createReview);
 app.post("/getFullTextSearch", queries.Reviews.getFullTextSearch);
+app.patch("/reviews/updateReview", jwtCheck, queries.Reviews.editReview);
 app.patch("/reviews/:reviewId/like", jwtCheck, queries.Reviews.likeReview);
 app.patch("/reviews/:reviewId/rate", jwtCheck, queries.Reviews.rateReview);
 app.patch(
@@ -50,6 +51,7 @@ app.patch(
   jwtCheck,
   queries.Reviews.addCommentToReview
 );
+app.delete("/reviews/:reviewId/delete", jwtCheck, queries.Reviews.deleteReview);
 app.delete("/deleteCategory", jwtCheck, queries.Reviews.deleteCategory);
 
 //media upload api
